Add BotJoinedChannel, clean GitHubWebhookCreated
BotJoinedChannel is not done, but it is mostly there. Still need to do
the repository drop down.

Break up things associated with repo/channel management so they can
more easily be composed.

Get real bot name in event subscriptions a few places.

Use objects when creating Slack buttons.

Do not talk about channel/repo linking in GitHubWebhookCreated
handler.  It does not do that, it just allows the owner to add the bot
to a channel.  Once there, the bot can suggest linking.