Simplify image URL regex, uniquify image URLs
The old image extraction URL went to great lengths to ensure if
parentheses occurred within the URL, they matched.  This causes a
repetition of a repetition which led to catastrophic backtracking, see
https://www.regular-expressions.info/catastrophic.html for more
information.  Unfortunately, JavaScript regular expressions do not
support possessive quantifiers or atomist groups, so we had to modify
the regular expression to avoid the issue.  The result is that we no
longer whether parentheses in the URL match.  That said, I don't think
there's any requirement they do anyway.

While testing this I found that invalid markdown could lead to a
similar situation, so I added a test for that too.  The resulting
image link is not valid, but neither is the markdown!

Added de-duplication of image URLs in attachments.

Fixes #136